fix: add email login
This fix introduces an email "login" form that hides the login container and displays the survey form upon successful "login".

BREAKING CHANGE:
There was no email "login" functionality and room numbers did not update dynamically based on the room type selection. Now, users must log in with an email to access the survey, so the moderator knows who has submitted.

POSSIBLE ADD:
This can give the possibility to send the results of the forms to the user, or the rework the login system where they can get a code thought the email so they can log in.